Add the current project backend files
This is the "proof-of-concept". Together all of these files accomplish what I want them to but it is not practical for actual use without a frontend. Also, if I hope to turn this into a UI I need to make these into an API rather than just random functions that run together.